[PLAYER-5446] Move panels components
