+++ sdk/react/src/panels/AudioView/AudioView.styles.json
