+++ sdk/react/src/panels/DiscoveryPanel/DiscoveryPanel.styles.json
