+++ sdk/react/src/panels/EndScreen/EndScreen.styles.json
