+++ sdk/react/src/panels/ErrorScreen/ErrorScreen.styles.json
