+++ sdk/react/src/panels/MoreDetailsScreen/MoreDetailsScreen.styles.json
