+++ sdk/react/src/panels/MoreOptionScreen/MoreOptionScreen.styles.json
