+++ sdk/react/src/panels/PlaybackSpeedPanel/PlaybackSpeedPanel.styles.json
