+++ sdk/react/src/panels/StartScreen/StartScreen.styles.json
