+++ sdk/react/src/panels/VideoView/VideoView.styles.json
