+++ sdk/react/src/panels/VolumePanel/VolumePanel.styles.json
